backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "flask-jwt-extended>=4.7.1",
    "flask-migrate>=4.1.0",
    "flask-restful>=0.3.10",
    "flask-sqlalchemy>=3.1.1",
    "mariadb>=1.1.14",
    "python-dotenv>=1.2.1",
]
